requirements.txt
ecdsa==0.18.1
requests==2.31.0
python-telegram-bot==20.2
base58==2.1.0